requirements.txt
PyQt5>=5.15.0
google-api-python-client>=2.134.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3
